apps/api/requirements.txt
Django>=6.0
djangorestframework
drf-spectacular
psycopg[binary]
python-dotenv
django-allauth[socialaccount]
dj-rest-auth
requests
